+++ app/models/db_trigger_generator_sqlite.rb
@@ -160,14 +160,14 @@ class DbTriggerGeneratorSqlite < DbTriggerGeneratorBase
       Database.select_all("PRAGMA table_info(#{table.name})").each do |table_column|
         columns.each do |trigger_column|
           if table_column.name.upcase == trigger_column[:column_name].upcase
-            trigger_column[:type]     = table_column.type
-            trigger_column[:notnull]  = table_column.notnull
+            trigger_column[:data_type]  = table_column.type
+            trigger_column[:nullable]  = table_column.notnull == 0
           end
         end
       end
 
       columns.each do |trigger_column|
-        raise "Column #{trigger_column[:column_name]} does not exist in table #{@schema.name}.#{table.name}" if trigger_column[:type].nil?
+        raise "Column #{trigger_column[:column_name]} does not exist in table #{@schema.name}.#{table.name}" if trigger_column[:data_type].nil?
       end
     end
   end
@@ -202,6 +202,8 @@ class DbTriggerGeneratorSqlite < DbTriggerGeneratorBase
   end
 
   # Build trigger code from hash
+  # @param [Table] table table object
+  # @param [String] operation 'I', 'U', 'D'
   def build_trigger_body(table, operation)
     table_config    = @expected_triggers[table.name]
     trigger_config  = table_config[operation]
@@ -215,7 +217,7 @@ class DbTriggerGeneratorSqlite < DbTriggerGeneratorBase
 
     payload = String.new
     accessors.each do |accessor|
-      payload << "\"#{accessor}\": #{payload_json(trigger_config, accessor, operation: operation)}"
+      payload << "\"#{accessor}\": #{payload_json(table, trigger_config, accessor, operation: operation)}"
       payload << "," if accessors.length == 2 && accessor == 'old'
     end
 
@@ -234,16 +236,22 @@ END;"
   end
 
   # Build JSON payload for new or old values
+  # @param table [Table] table object
   # @param trigger_config [Hash] configuration for trigger { }
   # @param accessor [String | NilClass] 'new', 'old' or nil for initialization
   # @param operation [String] 'I', 'U', 'D' or 'i' for initialization
   # @return [String] SQL expression for JSON payload
-  def payload_json(trigger_config, accessor, operation:)
+  def payload_json(table, trigger_config, accessor, operation:)
+    columns = trigger_config[:columns]
+    if table.yn_payload_pkey_only == 'Y'                                        # only primary key columns in payload requested
+      columns = table.pkey_columns                                              # only primary key columns { column_name: 'COL1', data_type: 'NUMBER', nullable: false }, ...
+    end
+
     json = "{".dup
-    json << trigger_config[:columns].map{|c| "\"#{c[:column_name]}\": '||#{convert_col(c, accessor)}||'"}.join(",\n")
+    json << columns.map{|c| "\"#{c[:column_name]}\": '||#{convert_col(c, accessor)}||'"}.join(",\n")
     expressions_to_use = trigger_config[:column_expressions].select {|ce| place_expression_in_section?(ce[:sql], accessor, operation) }
     unless expressions_to_use.empty?
-      json << ",\n" unless trigger_config[:columns].empty?
+      json << ",\n" unless columns.empty?
       json << expressions_to_use
                 .sort_by{|ce| ce[:id]}
                 .map{|ce| prepare_column_expressions(ce[:sql], operation)}
@@ -290,7 +298,7 @@ END;"
 
     sql = "\
 INSERT INTO Event_Logs(Table_ID, Operation, DBUser, Created_At, Payload, Msg_Key, Transaction_ID)
-SELECT #{table.id}, 'i', 'main', strftime('%Y-%m-%d %H-%M-%f','now'), '\"new\": #{payload_json(trigger_config, nil, operation: 'i')}', #{message_key_sql(table, 'i')},
+SELECT #{table.id}, 'i', 'main', strftime('%Y-%m-%d %H-%M-%f','now'), '\"new\": #{payload_json(table, trigger_config, nil, operation: 'i')}', #{message_key_sql(table, 'i')},
         #{table.yn_record_txid == 'Y' ? "'Dummy Tx-ID'" : "NULL" }
 FROM   main.#{table.name}
 "
@@ -340,12 +348,11 @@ FROM   main.#{table.name}
       when 'i' then nil                                                         # initialization of table data
       end
 
-    pk_columns = Database.select_all("PRAGMA table_info(#{table.name})").select{|c| c.pk > 0}
-    raise "DbTriggerSqlite.message_key_sql: Table #{table_name} does not have any primary key column" if pk_columns.length == 0
+    raise "DbTriggerSqlite.message_key_sql: Table #{table_name} does not have any primary key column" if table.pkey_columns.empty?
 
     result = "'{'||".dup
-    result << pk_columns
-                .map{|pkc| "'\"#{pkc['name']}\": '||#{convert_col({column_name: pkc['name'], type: pkc['type']}, pk_accessor)}" }
+    result << table.pkey_columns
+                .map{|pkc| "'\"#{pkc[:column_name]}\": '||#{convert_col(pkc, pk_accessor)}" }
                 .join("||','||")
     result << "||'}'"
 
@@ -375,7 +382,7 @@ FROM   main.#{table.name}
   def old_new_compare(columns)
     columns.map{|c|
       result = "old.#{c[:column_name]} != new.#{c[:column_name]}"
-      if c[:notnull] == 0
+      if c[:nullable]
         result << " OR (old.#{c[:column_name]} IS NULL AND new.#{c[:column_name]} IS NOT NULL)"
         result << " OR (old.#{c[:column_name]} IS NOT NULL AND new.#{c[:column_name]} IS NULL)"
       end
@@ -409,13 +416,13 @@ FROM   main.#{table.name}
     local_accessor = accessor.nil? ? '' : "#{accessor}."
     col_expr = "#{local_accessor}#{column_hash[:column_name]}"
     result = String.new
-    result << "CASE WHEN #{col_expr} IS NULL THEN 'null' ELSE '\"'||#{col_expr}||'\"' END"        if column_hash[:type] == 'BLOB'
-    result << "CASE WHEN #{col_expr} IS NULL THEN 'null' ELSE '\"'||#{col_expr}||'\"' END"        if column_hash[:type] =~ /datetime/i
-    result << "CASE WHEN #{col_expr} IS NULL THEN 'null' ELSE #{col_expr} END"                    if column_hash[:type] =~ /number/i || column_hash[:type] =~ /int/i
-    if column_hash[:type] =~ /char/i || column_hash[:type] =~ /text/i || column_hash[:type] =~ /varchar/i || column_hash[:type] =~ /clob/i
+    result << "CASE WHEN #{col_expr} IS NULL THEN 'null' ELSE '\"'||#{col_expr}||'\"' END"        if column_hash[:data_type] == 'BLOB'
+    result << "CASE WHEN #{col_expr} IS NULL THEN 'null' ELSE '\"'||#{col_expr}||'\"' END"        if column_hash[:data_type] =~ /datetime/i
+    result << "CASE WHEN #{col_expr} IS NULL THEN 'null' ELSE #{col_expr} END"                    if column_hash[:data_type] =~ /number/i || column_hash[:data_type] =~ /int/i
+    if column_hash[:data_type] =~ /char/i || column_hash[:data_type] =~ /text/i || column_hash[:data_type] =~ /varchar/i || column_hash[:data_type] =~ /clob/i
       result << "CASE WHEN #{col_expr} IS NULL THEN 'null' ELSE '\"'||REPLACE(#{col_expr}, '\"', '\\\"')||'\"' END"
     end
-    raise "Unsupported data type '#{column_hash[:type]}'" if result.length == 0
+    raise "Unsupported data type '#{column_hash[:data_type]}'" if result.length == 0
     result
   end